refactor: 폴더 디렉터리 구조 & Layout 생성
