+++ src/components/Layout/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Footer.css'
+import '../css/Footer.css'
 
 const Footer = () => {
   return (
