+++ src/components/Layout/Layout.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import Header from './Header'
+import Footer from './Footer'
+
+const Layout = ({ children }) => {
+  return (
+    <>
+      <Header />
+        {children}
+      <Footer />
+    </>
+  )
+}
+
+export default Layout
