+++ src/pages/Home.jsx
@@ -1,22 +1,17 @@
-
-import Header from '../components/Header'
 import HeroSection from '../components/HeroSection'
 import DescriptionSection from '../components/DescriptionSection'
 import PopuplarTopicSection from '../components/PopuplarTopicSection'
 import ClosingSection from '../components/ClosingSection'
-import Footer from '../components/Footer'
-import './Home.css'
+import Layout from '../components/Layout/Layout';
 
 const Home = () => {
   return (
-    <div className='HomeContainer'>
-      <Header />
+    <Layout>
       <HeroSection />
       <DescriptionSection />
       <PopuplarTopicSection />
       <ClosingSection />
-      <Footer />
-    </div>
+    </Layout>
   )
 };
